Storing of best_x between evaluations
Since x is mostly an array, the assignment self.best_x = x is done as a
reference. Then, if the original array x is modified in a special way,
e.g. x[0] = 1, the stored value self.best_x is inherently changed too.
This behavior is unintended and its cause difficult to find.